+++ src/main/java/com/epam/aidial/cfg/client/DeploymentClient.java
@@ -1,5 +1,6 @@
 package com.epam.aidial.cfg.client;
 
+import com.epam.aidial.cfg.client.dto.ToolSetDataDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,11 +11,16 @@ import java.util.Map;
         name = "deploymentClient",
         url = "${core.client.url}",
         configuration = {
-                AuthorizationCoreClientConfiguration.class
+                AuthorizationCoreClientConfiguration.class,
+                RetryClientConfiguration.class,
+                FeignErrorDecoderConfiguration.class
         }
 )
 public interface DeploymentClient {
 
     @GetMapping("/v1/deployments/{deploymentName}/configuration")
     Map<String, Object> getConfiguration(@PathVariable String deploymentName);
+
+    @GetMapping(value = "/openai/toolsets/{toolSetName}")
+    ToolSetDataDto getToolSet(@PathVariable String toolSetName);
 }
