--- src/main/java/com/epam/aidial/cfg/client/OpenaiDeploymentsClient.java
@@ -1,21 +1,0 @@
-package com.epam.aidial.cfg.client;
-
-import com.epam.aidial.cfg.client.dto.ToolSetDataDto;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-
-@FeignClient(
-        name = "openaiDeploymentsClient",
-        url = "${core.client.url}",
-        configuration = {
-                AuthorizationCoreClientConfiguration.class,
-                RetryClientConfiguration.class,
-                FeignErrorDecoderConfiguration.class
-        })
-public interface OpenaiDeploymentsClient {
-
-    @GetMapping(value = "/openai/toolsets/{toolSetName}")
-    ToolSetDataDto getOpenaiToolSet(@PathVariable String toolSetName);
-
-}
