+++ src/test/java/com/epam/aidial/cfg/functional/config/FunctionalTestConfiguration.java
@@ -1,10 +1,10 @@
 package com.epam.aidial.cfg.functional.config;
 
 import com.epam.aidial.cfg.client.AnonymousCoreConfigClient;
-import com.epam.aidial.cfg.client.OpenaiDeploymentsClient;
+import com.epam.aidial.cfg.client.DeploymentClient;
 import com.epam.aidial.cfg.client.ResourceCredentialClient;
-import com.epam.aidial.cfg.client.mapper.OpenaiDeploymentsClientMapper;
-import com.epam.aidial.cfg.client.mapper.OpenaiDeploymentsClientMapperImpl;
+import com.epam.aidial.cfg.client.mapper.DeploymentClientMapper;
+import com.epam.aidial.cfg.client.mapper.DeploymentClientMapperImpl;
 import com.epam.aidial.cfg.client.mapper.ResourceCredentialClientMapper;
 import com.epam.aidial.cfg.client.mapper.ResourceCredentialClientMapperImpl;
 import com.epam.aidial.cfg.client.mcp.McpClientFactory;
@@ -40,7 +40,7 @@ import com.epam.aidial.cfg.features.flag.aspect.FeatureFlagGateEvaluationAspect;
 import com.epam.aidial.cfg.functional.tests.history.TestHistoryFacade;
 import com.epam.aidial.cfg.mapper.ResourceCredentialMapper;
 import com.epam.aidial.cfg.mapper.ResourceCredentialMapperImpl;
-import com.epam.aidial.cfg.service.OpenaiDeploymentsService;
+import com.epam.aidial.cfg.service.CoreDeploymentService;
 import com.epam.aidial.cfg.service.ResourceCredentialService;
 import com.epam.aidial.cfg.service.config.export.CoreConfigAggregatorService;
 import com.epam.aidial.cfg.service.config.reload.CoreConfigReloadCache;
@@ -187,18 +187,18 @@ public class FunctionalTestConfiguration {
     }
 
     @Bean
-    public OpenaiDeploymentsClientMapper openaiDeploymentsClientMapper() {
-        return new OpenaiDeploymentsClientMapperImpl();
+    public DeploymentClientMapper deploymentClientMapper() {
+        return new DeploymentClientMapperImpl();
     }
 
     @Bean
-    public OpenaiDeploymentsClient openaiDeploymentsClient() {
-        return Mockito.mock(OpenaiDeploymentsClient.class);
+    public DeploymentClient deploymentClient() {
+        return Mockito.mock(DeploymentClient.class);
     }
 
     @Bean
-    public OpenaiDeploymentsService openaiDeploymentsService(OpenaiDeploymentsClient client, OpenaiDeploymentsClientMapper mapper) {
-        return new OpenaiDeploymentsService(client, mapper);
+    public CoreDeploymentService coreDeploymentService(DeploymentClient client, DeploymentClientMapper mapper) {
+        return new CoreDeploymentService(client, mapper);
     }
 
     @Bean
